Update IrisClassification
rather than just printing all the data in the iris data set, the program now trains a DecisionTreeClassifier to classify each of the 3 types of iris flower. The classifier is tested with 1 example of each type of iris flower, which are removed from the training data prior to training the classifier.